+++ src/main/java/com/dfo/dunsee/search/repository/CharacterInfoRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface CharacterInfoRepository extends JpaRepository<CharacterInfo, Integer> {
+public interface CharacterInfoRepository extends JpaRepository<CharacterInfo, Long> {
 
   List<CharacterInfo> findByAdventureName(String adventureName);
